Day 1 lab/requirements.txt
pandas==2.2.3
matplotlib==3.9.4
numpy==1.26.4
joblib==1.4.2
scikit-learn==1.8.0
seaborn==0.13.2
pyyaml==6.0.2
torch==2.6.0
mediapipe==0.10.21
opencv-python==4.11.0.86
